+++ lib/repeatable/expression.rb
@@ -1,3 +1,4 @@
+# typed: strict
 module Repeatable
   module Expression
   end