+++ lib/repeatable/expression/biweekly.rb
@@ -1,23 +1,31 @@
+# typed: strict
 module Repeatable
   module Expression
     class Biweekly < Date
+      sig { params(weekday: Integer, start_after: Object).void }
       def initialize(weekday:, start_after: ::Date.today)
         @weekday = weekday
-        @start_after = Conversions::Date(start_after)
+        @start_after = T.let(Conversions::Date(start_after), ::Date)
+        @_first_occurrence = T.let(find_first_occurrence, ::Date)
       end
 
+      sig { override.params(date: ::Date).returns(T::Boolean) }
       def include?(date)
-        date >= start_after && (date - first_occurrence) % 14 == 0
+        date >= start_after && (date - _first_occurrence) % 14 == 0
       end
 
       private
 
-      attr_reader :weekday, :start_after
+      sig { returns(Integer) }
+      attr_reader :weekday
 
-      def first_occurrence
-        @first_occurrence ||= find_first_occurrence
-      end
+      sig { returns(::Date) }
+      attr_reader :start_after
 
+      sig { returns(::Date) }
+      attr_reader :_first_occurrence
+
+      sig { returns(::Date) }
       def find_first_occurrence
         days_away = weekday - start_after.wday
         days_away += 7 if days_away <= 0