+++ lib/repeatable/expression/day_in_month.rb
@@ -1,12 +1,15 @@
+# typed: strict
 module Repeatable
   module Expression
     class DayInMonth < Date
       include LastDateOfMonth
 
+      sig { params(day: Integer).void }
       def initialize(day:)
         @day = day
       end
 
+      sig { override.params(date: ::Date).returns(T::Boolean) }
       def include?(date)
         if day < 0
           date - last_date_of_month(date) - 1 == day
@@ -17,6 +20,7 @@ module Repeatable
 
       private
 
+      sig { returns(Integer) }
       attr_reader :day
     end
   end