+++ lib/repeatable/parse_error.rb
@@ -1,3 +1,4 @@
+# typed: strict
 module Repeatable
   class ParseError < StandardError
   end