+++ lib/repeatable/types.rb
@@ -1,0 +1,6 @@
+# typed: strict
+module Repeatable
+  module Types
+    SymbolHash = T.type_alias { T::Hash[Symbol, T.untyped] }
+  end
+end