+++ spec/repeatable/expression/base_spec.rb
@@ -1,3 +1,4 @@
+# typed: false
 require "spec_helper"
 
 module Repeatable
@@ -19,18 +20,6 @@ module Repeatable
             expect(described_class === Repeatable::Expression::Set.new).to eq(true)
             expect(described_class === "").to eq(false)
           end
-        end
-      end
-
-      describe "#include?" do
-        it "raises a NotImplemented Error" do
-          expect { subject.include?("anything") }.to raise_error(NotImplementedError)
-        end
-      end
-
-      describe "#to_h" do
-        it "raises a NotImplemented Error" do
-          expect { subject.to_h }.to raise_error(NotImplementedError)
         end
       end