+++ spec/support/expression_examples.rb
@@ -1,3 +1,4 @@
+# typed: false
 shared_examples "an expression" do
   describe "#include?" do
     it "must be implemented" do